requirements.txt
Django==3.0.7
djangorestframework==3.11.0
flake8==3.7.9
mypy==0.770
django-stubs==1.5.0
djangorestframework-stubs==1.2.0
typing-extensions==3.7.4.2
Pillow==7.2.0
django-ical==1.7.1
drf-yasg==1.17.1
djangorestframework-simplejwt==4.4.0
